+++ gamepatches.py
@@ -1400,12 +1400,19 @@ class GamePatcher:
         self.exe_root_path = exe_root_path
         self.actual_extract_path = actual_extract_path
         self.modified_extract_path = modified_extract_path
+
         self.patcher = AllPatcher(
             actual_extract_path=actual_extract_path,
             modified_extract_path=modified_extract_path,
             oarc_cache_path=oarc_cache_path,
             arc_replacement_path=arc_replacement_path,
             assets_path=RANDO_ROOT_PATH / "assets",
+            current_player_model_pack_name=self.placement_file.options[
+                "selected-player-model-pack"
+            ],
+            current_loftwing_model_pack_name=self.placement_file.options[
+                "selected-loftwing-model-pack"
+            ],
             copy_unmodified=False,
         )
         self.text_labels = {}